Перенос списка моковых предметов в отдельную папку utils, в отдельный файл
